requirements.txt
# Core API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Document Processing
pdfplumber==0.11.4
pymupdf==1.24.10
python-docx==1.1.2
openpyxl==3.1.5
Pillow==10.4.0

# Ollama (free local LLM)
httpx==0.27.2

# Data & Analytics
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.1
shap==0.46.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Document Generation
reportlab==4.2.2

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tenacity==9.0.0
python-multipart==0.0.9
